wait for scss

+++ recipe-app/src/components/common/RecipeCard.jsx
@@ -6,13 +6,14 @@ import './RecipeCard.scss';
 const RecipeCard = ({ recipe, onOpenDetails, onDelete, onToggleFavorite }) => {
     const handleButtonClick = (event) => {
         event.stopPropagation(); // 阻止事件冒泡
-        if (onDelete) {
-            onDelete(recipe);
-        }
-        if (onToggleFavorite) {
-            onToggleFavorite(recipe.id);
-        }
+        onDelete && onDelete(recipe);
     };
+
+    const handleFavoriteClick = (event) => {
+        event.stopPropagation(); // 阻止事件冒泡
+        onToggleFavorite && onToggleFavorite(recipe.id);
+    };
+
     return (
         <div className="recipe-card" onClick={() => onOpenDetails(recipe)}>
             <img src={`http://localhost:8000${recipe.imagePath}`} alt={recipe.name} />
@@ -26,10 +27,10 @@ const RecipeCard = ({ recipe, onOpenDetails, onDelete, onToggleFavorite }) => {
             )}
             {onToggleFavorite && (
                 <FavoriteButton 
-                isFavorited={recipe.isFavorited} 
-                onToggle={()=> onToggleFavorite(recipe.id)} />
-            )
-            }
+                    isFavorited={recipe.isFavorited} 
+                    onToggle={handleFavoriteClick} 
+                />
+            )}
         </div>
     );
 };
